printf prolect

+++ main.h
@@ -50,7 +50,8 @@ typedef struct converter_s
 {
 	unsigned char specifier;
 
-	unsigned int (*func)(va_list, buffer_t *, size_t char, int, int, size_t char);
+	unsigned int (*func)(va_list, buffer_t *, unsigned char
+			int, int, unsigned char);
 } converter_t;
 
 /**
